requirements.txt
pygame-ce
black
flake8
mypy
pytest
